requirements.txt
requests==2.31.0
pytest==7.4.3
python-dotenv==1.0.0
faker==22.0.0
jsonschema==4.20.0
pytest-html==4.1.1